calc min/max zoom on metadata read
