+++ lib/metadata.js
@@ -1,6 +1,9 @@
 var queue = require('queue-async');
 var extent = require('geojson-extent');
 var through = require('through2');
+var SphericalMercator = require('sphericalmercator');
+var merc = new SphericalMercator();
+var _ = require('lodash');
 
 var Metadata = module.exports = function(dyno, dataset) {
 
@@ -35,8 +38,7 @@ var Metadata = module.exports = function(dyno, dataset) {
     metadata.getInfo = function(callback) {
         dyno.getItem(key, function(err, item) {
             if (err) return callback(err);
-            info = item || {};
-            callback(null, info);
+            callback(null, item ? prepare(item) : {});
         });
     };
 
@@ -114,7 +116,7 @@ var Metadata = module.exports = function(dyno, dataset) {
             .on('error', callback)
             .on('finish', function() {
                 dyno.putItem(info, function(err) {
-                    callback(err, info);
+                    callback(err, prepare(info));
                 });
             });
     };
@@ -209,3 +211,33 @@ var Metadata = module.exports = function(dyno, dataset) {
 
     return metadata;
 };
+
+function prepare(info) {
+    var range = zoomRange(info.size, [info.west, info.south, info.east, info.west]);
+    var result = _.clone(info);
+    result.minzoom = range.min;
+    result.maxzoom = range.max;
+    return result;
+}
+
+function zoomRange(bytes, extent) {
+    var maxSize = 500 * 1024;
+    var maxzoom = 14;
+    for (z = 22; z >= 0; z--) {
+        var bounds = merc.xyz(extent, z, false, 4326);
+        var x = (bounds.maxX - bounds.minX) + 1;
+        var y = (bounds.maxY - bounds.minY) + 1;
+        var tiles = x * y;
+        var avgTileSize = bytes / tiles;
+
+        // The idea is that tilesize of ~1000 bytes is usually the most detail
+        // needed, and no need to process tiles with higher zoom
+        if (avgTileSize < 1000) maxzoom = z;
+
+        // Tiles are getting too large at current z
+        if (avgTileSize > maxSize) return { min: z, max: maxzoom };
+
+        // If all the data fits into one tile, it'll fit all the way to z0
+        if (tiles === 1 || z === 0) return { min: 0, max: maxzoom };
+    }
+}
