+++ package.json
@@ -37,6 +37,7 @@
     "lodash": "~2.4.1",
     "minimist": "0.0.9",
     "queue-async": "~1.0.7",
+    "sphericalmercator": "^1.0.3",
     "through2": "^2.0.0",
     "tilebelt": "^0.5.2",
     "uniq": "~1.0.0"
