+++ test/index.js
@@ -1,5 +1,6 @@
 require('./config');
 require('./indexing');
+require('./metadata');
 require('./bbox_query');
 require('./batch.test.js');
 require('./cli.test.js');
